commit 1
This is the data and python code for getting the Number of accident by climate condition.

Here's how I did it, as the dataset was bit incomplete, I looked for the full dataset of traffic collision for city of Seattle, and I found the dataset from their website. Then I try to match your dataset with the complete dataset based on OBJECTID column. After doing the matching using left outer join, I did the analysis on number accident based on the weather condition. hope it answer what you were looking for